use bigint

+++ db/models/archives.py
@@ -1,21 +1,21 @@
 from sqlalchemy import (
-    Column, String, Integer, DateTime, Text
+    Column, String, BigInteger, DateTime, Text
 )
 from sqlalchemy.schema import Index
 from db.models import Base
 from arrow import Arrow
 
+
 class Archive(Base):
     __tablename__ = 'archives'
 
-    id = Column(Integer, autoincrement=True, primary_key=True)
+    id = Column(BigInteger, autoincrement=True, primary_key=True)
     hash = Column(String, nullable=False, unique=True)
     published = Column(DateTime(timezone=True), nullable=False)
     title = Column(String, nullable=False)
     summary = Column(Text, nullable=False)
     link = Column(String, nullable=False)
     source = Column(String, nullable=False)
-
 
     def __repr__(self):
         return '<Archive(title="{}")>'.format(self.title)
